Fix LightSwitch tests: extract shared types, eliminate hollow tests
- Created LightSwitchTypes.h with ScheduleMode, ToString/FromString, LightSwitchConfig
- Modified LightSwitchSettings.h to include LightSwitchTypes.h
- Removed copy-pasted definitions from test file (20 tests now test real code)
- Extracted CoordinatesAreValid to LightSwitchUtils.h as free function
- Added 6 CoordinatesAreValid tests
- Added 2 polar-region SunCalc edge case tests (documenting -1 sentinel bug)
- Added deep comments to all 56 test methods
